Utilisation de probabilités avec des double entre 0 et 100 (pour éviter des problèmes insolubles) Création d'une vue de wizard pas terrible Amélioration de l'architecture des classes Amélioration de l'architecture des Event en utilisant des interfaces-tag

+++ group3_genetic/src/main/java/fr/isen/cir56/group3_genetic/Operator/AbstractCrossoverOperator.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public abstract class AbstractCrossoverOperator extends AbstractOperator {
 
-	public AbstractCrossoverOperator(float p) throws InvalidProbabilityValueException {
+	public AbstractCrossoverOperator(double p) throws InvalidProbabilityValueException {
 		super(p);
 	}
