+++ group3_genetic/src/main/java/fr/isen/cir56/group3_genetic/Operator/OperatorInterface.java
@@ -1,13 +1,19 @@
 package fr.isen.cir56.group3_genetic.Operator;
 
-import fr.isen.cir56.group3_genetic.Utils.Math.Probability.InvalidProbabilityValueException;
 import fr.isen.cir56.group3_genetic.PopulationInterface;
+import fr.isen.cir56.group3_genetic.Utils.Math.Probability.InvalidProbabilityValueException;
 
 
 public interface OperatorInterface {
 
 	void evaluate(PopulationInterface population);
-	void setProbability(float probability) throws InvalidProbabilityValueException;
-	float getProbability();
+	
+	/**
+	 * 
+	 * @param probability between 0.00 and 100.0
+	 * @throws InvalidProbabilityValueException 
+	 */
+	void setProbability(double probability) throws InvalidProbabilityValueException;
+	double getProbability();
 	
 }
