+++ group3_genetic/src/main/java/fr/isen/cir56/group3_genetic/Utils/Math/Probability/InvalidProbabilityValueException.java
@@ -5,9 +5,9 @@ package fr.isen.cir56.group3_genetic.Utils.Math.Probability;
  * @author Louis VICAINNE louis.vicainne@gmail.com
  */
 public class InvalidProbabilityValueException extends Exception {
-	public final float p;
+	public final double p;
 
-	public InvalidProbabilityValueException(float p) {
+	public InvalidProbabilityValueException(double p) {
 		this.p = p;
 	}
